- avoiding instability in biquad filter - temp file no longer deleted when playing song from project mgr - added warnings for Reaper features which break conversion:   - media item MIDI offset / start offset   - play rate != 1   - fade ins / outs on media items   - media item volume level - channel data removed from reaper MIDI events - fixing some build issues in 32-bit MinSizeRel

+++ WaveSabreCore/include/WaveSabreCore/Maj7.hpp
@@ -578,7 +578,7 @@ namespace WaveSabreCore
 					lfo.mPhase.BeginBlock();
 				}
 
-				for (size_t iv = 0; iv < mMaxVoices; ++ iv)
+				for (size_t iv = 0; iv < (size_t)mMaxVoices; ++ iv)
 				{
 					mMaj7Voice[iv]->BeginBlock();
 				}
@@ -589,8 +589,7 @@ namespace WaveSabreCore
 				{
 					float s[2] = { 0 };
 
-					//for (auto* v : mMaj7Voice)
-					for (size_t iv = 0; iv < mMaxVoices; ++iv)
+					for (size_t iv = 0; iv < (size_t)mMaxVoices; ++iv)
 					{
 						mMaj7Voice[iv]->ProcessAndMix(s);
 					}
